requirements.txt
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.1
gym==0.26.2
gym-notices==0.0.8
imageio==2.31.1
kiwisolver==1.4.4
lazy_loader==0.3
matplotlib==3.7.2
networkx==3.1
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
tifffile==2023.7.18
